+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/View/Camera/CameraTimerView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct CameraTimerView: View {
-    @StateObject var motionManager = MotionManager()
+    @StateObject var motionManager = MotionViewModel()
     @Binding var delayTime: Double
     @Binding var isPushed: Int
     var body: some View {
